+++ include/counter/counter-internal.h
@@ -15,12 +15,12 @@
 #include <counter/counter-types.h>
 #include <counter/config.h>
 
-static inline size_t lttng_counter_get_index(const struct lib_counter_config *config,
+static inline int64_t lttng_counter_get_index(const struct lib_counter_config *config,
 					     struct lib_counter *counter,
 					     int64_t *dimension_indexes)
 {
 	size_t nr_dimensions = counter->nr_dimensions, i;
-	size_t index = 0;
+	int64_t index = 0;
 
 	for (i = 0; i < nr_dimensions; i++) {
 		struct lib_counter_dimension *dimension = &counter->dimensions[i];
