+++ include/counter/counter.h
@@ -18,7 +18,7 @@
 /* max_nr_elem is for each dimension. */
 struct lib_counter *lttng_counter_create(const struct lib_counter_config *config,
 					 size_t nr_dimensions,
-					 const size_t *max_nr_elem,
+					 const int64_t *max_nr_elem,
 					 int64_t global_sum_step);
 void lttng_counter_destroy(struct lib_counter *counter);
